Add 3 new traps in the game class and unit test for all new traps
Added the 3 new traps + the steel trap to the game and added the new unit test for the game class and audioPlayer class. Names for the traps are Rat Trap, Cat Trap, A24 Trap(poison trap for rat and stoat) and the Steel Trap(General All-purpose trap). JSON and POOL files are updated with the new trap data.